backend/requirements.txt
Django==3.2.19
django-cors-headers==4.0.0
django-extra-fields==3.0.2
djangorestframework==3.14.0
django-filter==23.2
djoser==2.2.0
Pillow==9.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
